Remove OpenShift-specific bundle component
Since the OCM/RBAC integration is now in production where OCM is using ConsoleDot
RBAC to manage users, groups and roles, we should remove the messaging that
OpenShift Clusters permissions are not managed with User Access, since they can
be now.

This change is similar to what's used in the Ansible bundle, defaulting to the
RHEL implementation of `CommonBundleView`.