fix: refine the invite users error messages
If a user has already been invited, but has not accepted the invitation,
the error message simply says, "Failed to invite user." This messaging
is unhelpful in that it doesn't inform the user that the intended
recipient has already been invited into the org.

This change notifies the user that the email already exists in the
database which should help them have a more clear understanding of _why_
the invitation failed.

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>